Move button to appkit
Moves the basic styled button, as well as
its styling method, to appkit.
