+++ frontend/src/routes/Upload.tsx
@@ -13,7 +13,8 @@ import { makeRoute } from "../rauta";
 import { ErrorDisplay, errorDisplayInfo } from "../util/err";
 import { useNavBlocker } from "./util";
 import CONFIG from "../config";
-import { Button, LinkButton } from "../ui/Button";
+import { Button } from "@opencast/appkit";
+import { LinkButton } from "../ui/LinkButton";
 import { boxError, ErrorBox } from "../ui/error";
 import { Form } from "../ui/Form";
 import { Input, TextArea } from "../ui/Input";
