+++ frontend/src/routes/manage/Realm/Content/Edit/EditMode/index.tsx
@@ -5,7 +5,7 @@ import { useForm, useFormContext, FormProvider } from "react-hook-form";
 import { bug, match } from "@opencast/appkit";
 
 import { ConfirmationModal, ConfirmationModalHandle } from "../../../../../../ui/Modal";
-import { Button } from "../../../../../../ui/Button";
+import { Button } from "@opencast/appkit";
 import { currentRef, useOnOutsideClick } from "../../../../../../util";
 import type { EditModeRealmData$key } from "./__generated__/EditModeRealmData.graphql";
 import type { EditModeFormRealmData$key } from "./__generated__/EditModeFormRealmData.graphql";
