+++ frontend/src/ui/InitialConsent.tsx
@@ -2,11 +2,9 @@ import { useRef } from "react";
 import CONFIG from "../config";
 import { currentRef, getTranslatedString, translatedConfig } from "../util";
 import { ModalHandle, Modal } from "@opencast/appkit";
-import { Button } from "./Button";
 import { TextBlock } from "./Blocks/Text";
 import { useTranslation } from "react-i18next";
-import { notNullish } from "@opencast/appkit";
-
+import { Button, notNullish } from "@opencast/appkit";
 
 const LOCAL_STORAGE_KEY = "tobiraUserConsent";
